Explicitly use version 3 of the API
Veersion 4 of the oVirt engine will support two versions of the API:
version 3, which will be backwards compatible with older versions of the
engine and version 4, which won't be compatible. The default used by the
server will be version 4, but clients can explicitly request a version
of the API using the "Version" header or the "/v3" URL prefix. The
header is the preferred mechanism because it is just ignored by older
versions of the server. This patch modifies the the "check_rhev3.pl"
script so that it always send the header explicitly, so that it will
work with version 4 of the engine.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>